LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/back-already-detached-expected.txt



FAIL back() in a detached window assert_equals: expected 2 but got 1
